tf-idf

+++ java/label/src/main/java/com/codeu/team34/label/WikiSearch.java
@@ -23,14 +23,14 @@ import redis.clients.jedis.Jedis;
 public class WikiSearch {
 	
 	// map from URLs that contain the term(s) to relevance score
-	private Map<String, Integer> map;
+	private Map<String, Double> map;
 
 	/**
 	 * Constructor.
 	 * 
 	 * @param map
 	 */
-	public WikiSearch(Map<String, Integer> map) {
+	public WikiSearch(Map<String, Double> map) {
 		this.map = map;
 	}
 	
@@ -40,8 +40,8 @@ public class WikiSearch {
 	 * @param url
 	 * @return
 	 */
-	public Integer getRelevance(String url) {
-		Integer relevance = map.get(url);
+	public Double getRelevance(String url) {
+		Double relevance = map.get(url);
 		return relevance==null ? 0: relevance;
 	}
 	
@@ -51,9 +51,9 @@ public class WikiSearch {
 	 * @param map
 	 */
 	public void print() {
-		List<Entry<String, Integer>> entries = sort();
+		List<Entry<String, Double>> entries = sort();
 		if(entries.size()>0){
-		for (Entry<String, Integer> entry: entries) {
+		for (Entry<String, Double> entry: entries) {
 			System.out.println(entry);
 		}
 		}else{
@@ -68,9 +68,9 @@ public class WikiSearch {
 	 * @return New WikiSearch object.
 	 */
 	public WikiSearch or(WikiSearch that) {
-		Map<String, Integer> union = new HashMap<String, Integer>(map);
+		Map<String, Double> union = new HashMap<String, Double>(map);
 		for (String term: that.map.keySet()) {
-			int relevance = totalRelevance(this.getRelevance(term), that.getRelevance(term));
+			double relevance = totalRelevance(this.getRelevance(term), that.getRelevance(term));
 			union.put(term, relevance);
 		}
 		return new WikiSearch(union);
@@ -83,10 +83,10 @@ public class WikiSearch {
 	 * @return New WikiSearch object.
 	 */
 	public WikiSearch and(WikiSearch that) {
-		Map<String, Integer> intersection = new HashMap<String, Integer>();
+		Map<String, Double> intersection = new HashMap<String, Double>();
 		for (String term: map.keySet()) {
 			if (that.map.containsKey(term)) {
-				int relevance = totalRelevance(this.map.get(term), that.map.get(term));
+				double relevance = totalRelevance(this.map.get(term), that.map.get(term));
 				intersection.put(term, relevance);
 			}
 		}
@@ -100,7 +100,7 @@ public class WikiSearch {
 	 * @return New WikiSearch object.
 	 */
 	public WikiSearch minus(WikiSearch that) {
-		Map<String, Integer> difference = new HashMap<String, Integer>(map);
+		Map<String, Double> difference = new HashMap<String, Double>(map);
 		for (String term: that.map.keySet()) {
 			difference.remove(term);
 		}
@@ -114,7 +114,7 @@ public class WikiSearch {
 	 * @param rel2: relevance score for the second search
 	 * @return
 	 */
-	protected int totalRelevance(Integer rel1, Integer rel2) {
+	protected double totalRelevance(Double rel1, Double rel2) {
 		// simple starting place: relevance is the sum of the term frequencies.
 		return rel1 + rel2;
 	}
@@ -124,17 +124,17 @@ public class WikiSearch {
 	 * 
 	 * @return List of entries with URL and relevance.
 	 */
-	public List<Entry<String, Integer>> sort() {
+	public List<Entry<String, Double>> sort() {
 		// NOTE: this can be done more concisely in Java 8.  See
 		// http://stackoverflow.com/questions/109383/sort-a-mapkey-value-by-values-java
 
 		// make a list of entries
-		List<Entry<String, Integer>> entries = 
-				new LinkedList<Entry<String, Integer>>(map.entrySet());
+		List<Entry<String, Double>> entries = 
+				new LinkedList<Entry<String, Double>>(map.entrySet());
 		
 		// make a Comparator object for sorting
-		Comparator<Entry<String, Integer>> comparator = new Comparator<Entry<String, Integer>>() {
-            public int compare(Entry<String, Integer> e1, Entry<String, Integer> e2) {
+		Comparator<Entry<String, Double>> comparator = new Comparator<Entry<String, Double>>() {
+            public int compare(Entry<String, Double> e1, Entry<String, Double> e2) {
                 //return e1.getValue().compareTo(e2.getValue());
             	//Make the order to descending order
             	return e2.getValue().compareTo(e1.getValue());
@@ -155,7 +155,7 @@ public class WikiSearch {
 	 * @return
 	 */
 	public static WikiSearch search(String term, JedisIndex index) {
-		Map<String, Integer> map = index.getCounts(term);
+		Map<String, Double> map = index.getCounts(term);
 		return new WikiSearch(map);
 	}
